+++ backend/app/models/audit_log.py
@@ -1,7 +1,7 @@
 """审计日志模型"""
 from datetime import datetime
 from typing import Optional
-from sqlalchemy import String, Integer, DateTime, Text, ForeignKey, BigInteger
+from sqlalchemy import String, Integer, DateTime, Text, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.sql import func
 from app.db.base import Base
@@ -9,7 +9,7 @@ from app.db.base import Base
 class AuditLog(Base):
     __tablename__ = "audit_logs"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[Optional[int]] = mapped_column(Integer, ForeignKey("users.id", ondelete="SET NULL"), nullable=True, comment="操作用户ID")
     username: Mapped[str] = mapped_column(String(64), nullable=False, comment="操作用户名")
     action: Mapped[str] = mapped_column(String(64), nullable=False, comment="操作类型：login/logout/create/update/delete/export")
